refs #56 - Isolate DI
